requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.13
uuid==1.30
pydantic==1.10.16
psycopg2-binary==2.9.6
python-multipart==0.0.6
click==8.1.7
h11==0.14.0
Pillow==10.0.0
starlette==0.26.1
typing_extensions==4.6.2
anyio==3.6.2
watchdog==2.1.9
fitz==1.19.6
pandas==1.3.3
numpy==1.21.2
sentence-transformers==2.2.2
faiss-cpu==1.7.1
transformers==4.9.2